rearranging code, properly deallocating output tensors, small script improvements..

+++ revert.sh
@@ -6,9 +6,28 @@ DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 cd "$DIR"
 
 #All OpenGL/BVH code relies on RGBDAcquisition code so we need to sync this as well
-cd RGBDAcquisition
+
+
+if [ -f dependencies/RGBDAcquisition/README.md ]; then
+echo "RGBDAcquisition appears to exist, reverting it .."
+cd dependencies/RGBDAcquisition
 git reset --hard HEAD
 git pull origin master
+cd "$DIR"
+else
+echo "Could not find RGBDAcquisition, please rerun the initialize.sh script .." 
+fi
+
+
+
+if [ -f dependencies/AmmarServer/README.md ]; then
+echo "AmmarServer appears to exist, reverting it .."
+cd dependencies/AmmarServer
+git reset --hard HEAD
+git pull origin master
+cd "$DIR"
+fi
+
 
 #Now sync rest of code
 cd "$DIR"
